added sample project with limited functionality for faster development iterations as well as functionality demos
