+++ My GPT Dashboard/My GPT Dashboard/Assets.xcassets/AccentColor.colorset/Contents.json
