+++ My GPT Dashboard/My GPT Dashboard/Assets.xcassets/AppIcon.appiconset/Contents.json
